etl/simple_project/app/requirements.txt
django==4.1.2
flake8==5.0.4
python-dotenv==0.21.0
django-split-settings==1.2.0
django-debug-toolbar==3.7.0
oitnb==0.2.2
isort==5.10.1
pytest==7.2.0
pydantic==1.10.2
asgiref==3.5.2
click==8.1.3
colorama==0.4.6
dacite==1.6.0
django-cors-headers==3.13.0
django-extensions==3.2.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
protobuf==4.21.9
psycopg2-binary==2.9.5
split-settings==1.0.0
sqlparse==0.4.3
tzdata==2022.6
uWSGI==2.0.21
zipp==3.10.0